+++ tropic01_model/main.c
@@ -7,6 +7,7 @@
 
 #include <arpa/inet.h>
 #include <errno.h>
+#include <stdio.h>
 #include <string.h>
 #include <unistd.h>